+++ cbits/bytes.c
@@ -36,9 +36,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #if defined(__x86_64__)
 #include <x86intrin.h>
 #endif
-#if defined(Z_USE_AVX512)
+#if defined(__AVX512BW__)
 #include <fastavx512bwbase64.h>
-#elif defined(Z_USE_AVX2)
+#elif defined(__AVX2__)
 #include <fastavxbase64.h>
 #endif
 
@@ -293,9 +293,9 @@ HsInt hs_hex_decode_ws(uint8_t* output, const uint8_t* input, HsInt input_off, H
 // Base64 codec
 
 void hs_base64_encode(char* output, HsInt output_off, const char* input, HsInt off, HsInt len){
-#if defined(Z_USE_AVX512)
+#if defined(__AVX512BW__)
     fast_avx512bw_base64_encode(output+output_off, input+off, len);
-#elif defined(Z_USE_AVX2)
+#elif defined(__AVX2__)
     fast_avx2_base64_encode(output+output_off, input+off, len);
 #else
     chromium_base64_encode(output+output_off, input+off, len);
@@ -303,9 +303,9 @@ void hs_base64_encode(char* output, HsInt output_off, const char* input, HsInt o
 }
 
 HsInt hs_base64_decode(char* output, const char* input, HsInt off, HsInt len){
-#if defined(Z_USE_AVX512)
+#if defined(__AVX512BW__)
     size_t r = fast_avx512bw_base64_decode(output, input+off, len);
-#elif defined(Z_USE_AVX2)
+#elif defined(__AVX2__)
     size_t r = fast_avx2_base64_decode(output, input+off, len);
 #else
     size_t r = chromium_base64_decode(output, input+off, len);
@@ -352,7 +352,7 @@ void hs_intersperse(unsigned char *q,
 
 // this function should be called with large enough len, otherwise underflow will happen.
 size_t hs_count_simd(unsigned char *str, size_t len, unsigned char w) {
-#if defined(Z_USE_AVX2) 
+#if defined(__AVX2__) 
     __m256i pat = _mm256_set1_epi8(w);
     size_t prefix = 0, res = 0;
     size_t i = 0;
@@ -457,7 +457,7 @@ size_t hs_count_simd(unsigned char *str, size_t len, unsigned char w) {
 }
 
 HsInt hs_count_ba(unsigned char *str, HsInt off, HsInt len, unsigned char w) {
-#if defined(__SSE4_2__) || defined(Z_USE_AVX2) || defined(Z_USE_AVX512)
+#if defined(__SSE4_2__) || defined(__AVX2__)
     if (len >= 1024) return (HsInt)hs_count_simd(str+off, len, w);
     else {
 #endif
@@ -466,7 +466,7 @@ HsInt hs_count_ba(unsigned char *str, HsInt off, HsInt len, unsigned char w) {
         for (res = 0; len-- != 0; ++str)
             res += *str == w;
         return res;
-#if defined(__SSE4_2__) || defined(Z_USE_AVX2) || defined(Z_USE_AVX512)
+#if defined(__SSE4_2__) || defined(__AVX2__)
     }
 #endif
 }