+++ cbits/text.c
@@ -30,7 +30,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <utf8rewind.h>
 #include <codepoint.h>
 
-#if defined(Z_USE_AVX512)
+#if defined(__AVX512F__) && defined(__AVX512VL__) && defined(__AVX512VBMI__)
 #include <simdasciicheck_avx512.h>
 #include <simdutf8check_avx512.h>
 #else
@@ -40,9 +40,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 HsInt ascii_validate(const char* p, HsInt off, HsInt len){
     const char* q = p + off;
-#if defined(Z_USE_AVX512)
+#if defined(AVX512_IMPLEMENTATION)
     return (HsInt)validate_ascii_fast_avx512(q, (size_t)len);
-#elif defined(Z_USE_AVX2)
+#elif defined(__AVX2__)
     return (HsInt)validate_ascii_fast_avx(q, (size_t)len);
 #elif defined(__SSE2__)
     return (HsInt)validate_ascii_fast(q, (size_t)len);
@@ -53,9 +53,9 @@ HsInt ascii_validate(const char* p, HsInt off, HsInt len){
 // for some reason unknown, on windows we have to supply a seperated version of ascii_validate
 // otherwise we got segfault if we import the same FFI with different type (Addr# vs ByteArray#)
 HsInt ascii_validate_addr(const char* p, HsInt len){
-#if defined(Z_USE_AVX512)
+#if defined(AVX512_IMPLEMENTATION)
     return (HsInt)validate_ascii_fast_avx512(p, (size_t)len);
-#elif defined(Z_USE_AVX2)
+#elif defined(__AVX2__)
     return (HsInt)validate_ascii_fast_avx(p, (size_t)len);
 #elif defined(__SSE2__)
     return (HsInt)validate_ascii_fast(p, (size_t)len);
@@ -66,9 +66,9 @@ HsInt ascii_validate_addr(const char* p, HsInt len){
 
 HsInt utf8_validate(const char* p, HsInt off, HsInt len){
     const char* q = p + off;
-#if defined(Z_USE_AVX512)
+#if defined(AVX512_IMPLEMENTATION)
     return (HsInt)validate_utf8_fast_avx512(q, (size_t)len);
-#elif defined(Z_USE_AVX2)
+#elif defined(__AVX2__)
     return (HsInt)validate_utf8_fast_avx(q, (size_t)len);
 #elif defined(__SSE2__)
     return (HsInt)validate_utf8_fast(q, (size_t)len);
@@ -79,9 +79,9 @@ HsInt utf8_validate(const char* p, HsInt off, HsInt len){
 // for some reason unknown, on windows we have to supply a seperated version of utf8_validate
 // otherwise we got segfault if we import the same FFI with different type (Addr# vs ByteArray#)
 HsInt utf8_validate_addr(const char* p, HsInt len){
-#if defined(Z_USE_AVX512)
+#if defined(AVX512_IMPLEMENTATION)
     return (HsInt)validate_utf8_fast_avx512(p, (size_t)len);
-#elif defined(Z_USE_AVX2)
+#elif defined(__AVX2__)
     return (HsInt)validate_utf8_fast_avx(p, (size_t)len);
 #elif defined(__SSE2__)
     return (HsInt)validate_utf8_fast(p, (size_t)len);